refactor: transfer send-join-us-form-fx to api folder
