--- src/@/project/effects/index.ts
@@ -1,1 +1,0 @@
-export { sendJoinUsFormFx } from './send-join-us-form-fx';
